+++ changelog.md
@@ -138,3 +138,5 @@
 
 - UI : General fixes on weapon details.
 - UI : Font on favorite table is now compliant with font policy.
+- UI : Some fixes are made
+- Colors : New combinaison.